fix: 소숫점 계산 방식 변경

+++ src/main/kotlin/lottery/domain/LottoResult.kt
@@ -1,6 +1,8 @@
 package lottery.domain
 
+import java.math.BigDecimal
+
 data class LottoResult(
-    val yield: Double,
+    val yield: BigDecimal,
     val statistics: Map<Rank, Int>
 )
